Created brandsRouter
The brandsRouter is an Express router that handles routes related to brand operations. It includes routes for creating, retrieving, updating, and deleting brands. These routes are associated with corresponding controller functions to perform the necessary actions. Additionally, the isLoggedIn middleware is used to ensure that certain routes are only accessible to authenticated users.